Implement a ProjectAwareTrait
