+++ src/App/Groups/Controller/Groups.php
@@ -50,7 +50,7 @@ class Groups extends AbstractTrackerController
 
 		$this->getContainer()->get('app')->getUser()->authorize('manage');
 
-			$this->model->setProject($this->getContainer()->get('app')->getProject());
+		$this->model->setProject($this->getContainer()->get('app')->getProject());
 		$this->view->setProject($this->getContainer()->get('app')->getProject());
 
 		return $this;
