+++ src/App/Tracker/Model/IssuesModel.php
@@ -8,8 +8,6 @@
 
 namespace App\Tracker\Model;
 
-use App\Projects\TrackerProject;
-
 use Joomla\Database\DatabaseQuery;
 use Joomla\String\StringHelper;
 
@@ -30,48 +28,6 @@ class IssuesModel extends AbstractTrackerListModel
 	 * @since  1.0
 	 */
 	protected $context = 'tracker.issues';
-
-	/**
-	 * Project object
-	 *
-	 * @var    TrackerProject
-	 * @since  1.0
-	 */
-	protected $project = null;
-
-	/**
-	 * Get the project.
-	 *
-	 * @return  \App\Projects\TrackerProject
-	 *
-	 * @since   1.0
-	 * @throws  \RuntimeException
-	 */
-	public function getProject()
-	{
-		if (is_null($this->project))
-		{
-			throw new \RuntimeException('Project not set');
-		}
-
-		return $this->project;
-	}
-
-	/**
-	 * Set the project.
-	 *
-	 * @param   TrackerProject  $project  The project.
-	 *
-	 * @return  $this  Method supports chaining
-	 *
-	 * @since   1.0
-	 */
-	public function setProject(TrackerProject $project)
-	{
-		$this->project = $project;
-
-		return $this;
-	}
 
 	/**
 	 * Method to get a DatabaseQuery object for retrieving the data set from a database.
